Convert it to a more general instrument "Fan" instead of "DomeFan". This way, telescope, camera and other instruments can also have fans.
